akencoder: separate audio rendering feature from producer.cpp
